Jeongjune

+++ dasony/src/main/java/com/ds/dasony/game/model/vo/Game.java
@@ -12,6 +12,6 @@ public class Game {
 	private String gameStatus;
 	private String pointStatus;
 	private String ticketStatus;
-	private int userNo;
+	private long userNo;
 	
 }
